% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected.R
\name{expected}
\alias{expected}
\title{Expected Outcome}
\usage{
expected(R_a, R_b, s = 400)
}
\arguments{
\item{R_a}{numeric vector}

\item{R_b}{numeric vector}

\item{s}{numeric vector}
}
\value{
expected outcome of game
}
\description{
Expected Outcome
}
\examples{
R_a <- 1600
R_b <- 1400
expected(R_a,R_b)
}
